refactor: Extracted gen_fn_info to file_util module.
