+++ vistock/plotly/pbv4s.py
@@ -15,6 +15,7 @@ import plotly.graph_objs as go
 from plotly.subplots import make_subplots
 
 from .. import tw
+from .. import file_util
 from . import fig_util as futil
 
 
@@ -160,7 +161,7 @@ def plot(symbol='TSLA', period='12mo', interval='1d',
     fig.show()
 
     # Write the figure to an HTML file
-    fn = futil.gen_fn_info(symbol, interval, df.index.values[-1], __file__)
+    fn = file_util.gen_fn_info(symbol, interval, df.index.values[-1], __file__)
     fig.write_html(f'{fn}.html')
